Remove LocaleUtil.getLocale and move its logic to LocaleProvider.getLocale
COAND-848
