+++ action-core/src/main/java/com/adyen/checkout/action/core/internal/ActionHandlingPaymentMethodConfigurationBuilder.kt
@@ -15,7 +15,6 @@ import com.adyen.checkout.await.AwaitConfiguration
 import com.adyen.checkout.components.core.internal.BaseConfigurationBuilder
 import com.adyen.checkout.components.core.internal.Configuration
 import com.adyen.checkout.core.Environment
-import com.adyen.checkout.core.internal.util.LocaleUtil
 import com.adyen.checkout.qrcode.QRCodeConfiguration
 import com.adyen.checkout.redirect.RedirectConfiguration
 import com.adyen.checkout.voucher.VoucherConfiguration
@@ -78,11 +77,12 @@ constructor(
      * @param clientKey Your Client Key used for internal network calls from the SDK to Adyen.
      */
     constructor(
+        @Suppress("unused")
         context: Context,
         environment: Environment,
         clientKey: String
     ) : this(
-        LocaleUtil.getLocale(context),
+        null,
         environment,
         clientKey,
     )
