+++ checkout-core/src/main/java/com/adyen/checkout/core/internal/util/LocaleProvider.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2024 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by josephj on 9/2/2024.
+ */
+
+package com.adyen.checkout.core.internal.util
+
+import android.content.Context
+import android.os.Build
+import androidx.annotation.RestrictTo
+import java.util.Locale
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+class LocaleProvider {
+
+    fun getLocale(context: Context): Locale {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+            context.resources.configuration.locales[0]
+        } else {
+            @Suppress("DEPRECATION")
+            context.resources.configuration.locale
+        }
+    }
+}
