+++ checkout-core/src/main/java/com/adyen/checkout/core/internal/util/LocaleUtil.kt
@@ -7,8 +7,6 @@
  */
 package com.adyen.checkout.core.internal.util
 
-import android.content.Context
-import android.os.Build
 import androidx.annotation.RestrictTo
 import java.util.IllformedLocaleException
 import java.util.Locale
@@ -18,21 +16,6 @@ import java.util.Locale
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 object LocaleUtil {
-
-    /**
-     * Get the current user Locale.
-     * @param context The context
-     * @return The user Locale
-     */
-    @JvmStatic
-    fun getLocale(context: Context): Locale {
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            context.resources.configuration.locales[0]
-        } else {
-            @Suppress("DEPRECATION")
-            context.resources.configuration.locale
-        }
-    }
 
     /**
      * Gets the language tag from a Locale.
